+++ paper/__init__.py
@@ -1,0 +1,2 @@
+# ruff: noqa: CPY001
+"""Executable paper, reproduction tools, and supporting evidence."""
