+++ paper/evidence/README.md
@@ -4,6 +4,6 @@ Store concise, reviewable evidence for targets from `../targets.tsv` here.
 
 Prefer one Markdown file per logical target group, such as `figure-2.md` or `toxcast-classifiers.md`.
 Include the target IDs, input commit, commands or notebook cells used, observed values, acceptance decision, and deviations.
-Link existing artifacts or the top-level `REPRODUCING.md` when they already provide sufficient evidence.
+Link existing artifacts or `paper/REPRODUCING.md` when they already provide sufficient evidence.
 
 Large regenerated datasets, caches, and temporary notebook outputs do not belong here.
