+++ paper/reproduction/report.json
@@ -2967,7 +2967,7 @@
       ]
     }
   },
-  "ledger_sha256": "e0bc966efd96e283084c386e6a209eb5580e56b0493572882d1c28657bab02e1",
+  "ledger_sha256": "5e59e1169788549c6ac38f4385648d38d6a8c77fa28b29cd4f7d9b9f31ae0b99",
   "mode": "tracked",
   "schema_version": 3,
   "summary": {
@@ -5250,7 +5250,7 @@
       ],
       "description": "Cytotoxicity points of departure",
       "deviation": "Published workbook keys match all 796 committed semantic rows and POD values agree to spreadsheet precision; regenerated gates pass with documented drift",
-      "evidence": "evidence/published-table-bridge.md; ../REPRODUCING.md",
+      "evidence": "evidence/published-table-bridge.md; REPRODUCING.md",
       "evidence_strength": "source-recomputation",
       "execution_outcome": "checked",
       "expected": "Published MT, cell-count, and LDH POD results",
@@ -5299,7 +5299,7 @@
       ],
       "description": "Cell Painting points of departure",
       "deviation": "Published workbook keys match all 10935 nonblank committed semantic rows and POD values agree to spreadsheet precision; regenerated gates pass with documented drift",
-      "evidence": "evidence/published-table-bridge.md; ../REPRODUCING.md",
+      "evidence": "evidence/published-table-bridge.md; REPRODUCING.md",
       "evidence_strength": "source-recomputation",
       "execution_outcome": "checked",
       "expected": "Published CellProfiler, CP-CNN, and DINO POD results",
@@ -5348,7 +5348,7 @@
       ],
       "description": "Activity summary across all assays",
       "deviation": "Published workbook matches all 1086 committed hit-summary rows; regenerated keys are exact and 1070 of 1086 complete hit calls agree",
-      "evidence": "evidence/published-table-bridge.md; ../REPRODUCING.md",
+      "evidence": "evidence/published-table-bridge.md; REPRODUCING.md",
       "evidence_strength": "source-recomputation",
       "execution_outcome": "checked",
       "expected": "Published cross-assay compound activity summary including the 121-compound all-POD subset",
@@ -5531,7 +5531,7 @@
       ],
       "description": "Concentration-response model and POD selection",
       "deviation": "Morphology selects minimum residual SD; cell count, MT, and LDH use scoresPOD default rounded-AIC selection despite the paper's general residual-SD wording",
-      "evidence": "evidence/method-deviations.md; ../REPRODUCING.md",
+      "evidence": "evidence/method-deviations.md; REPRODUCING.md",
       "evidence_strength": "documentary-trace",
       "execution_outcome": "documentary-only",
       "expected": "Paper describes eight named models, residual-SD selection, DMSO 95th-percentile benchmark, and confidence-ratio and concentration filters",
@@ -5991,7 +5991,7 @@
     "2_downstream_analysis/manuscript_notebooks/3_1_toxcast_endpoints.ipynb",
     "2_downstream_analysis/other_notebooks/01_checkwelleffects.ipynb",
     "README.md",
-    "REPRODUCING.md",
+    "paper/REPRODUCING.md",
     "paper/evidence/activity-and-figure-2a.md",
     "paper/evidence/bioactivity-and-figure-2b-d.md",
     "paper/evidence/experimental-design-and-figure-1.md",
