+++ paper/tests/__init__.py
@@ -1,0 +1,2 @@
+# ruff: noqa: CPY001
+"""Tests for the executable paper and reproduction tools."""
