+++ paper/tests/test_render_sfig1.py
